requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.0.1
future==0.18.2
idna==2.8
lxml==4.4.2
Pillow==6.2.1
pycparser==2.19
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.7
xmltodict==0.12.0
